some adjustments to the copybara sync
PiperOrigin-RevId: 364560336
