+++ kubric/datasets/klevr_uni.py
@@ -1,3 +1,17 @@
+# Copyright 2021 The Kubric Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 # Copyright 2021 The Kubric Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -182,7 +196,7 @@ class Klevr(tfds.core.BeamBasedBuilder):
     """Returns SplitGenerators."""
     del unused_dl_manager
     path = tfds.core.as_path('gs://research-brain-kubric-xgcp/jobs/klevr_uni_v0_10k',)
-    all_subdirs = sorted([d for d in path.glob('*')], key=lambda x: int(x.name))
+    all_subdirs = sorted(list(path.glob('*')), key=lambda x: int(x.name))
     logging.info('Found %d sub-folders in master path: %s', len(all_subdirs), path)
 
     str_directories = [str(d) for d in all_subdirs]
